Combine two scripts into one
This removes the need to import the vectors twice, once per script, which takes time
